+++ traffic_emissions/components/utils.py
@@ -1,3 +1,5 @@
+from enum import Enum
+
 import pandas as pd
 from climatoology.base.artifact import ContinuousLegendData
 from matplotlib import colors
@@ -6,6 +8,28 @@ from pydantic_extra_types.color import Color
 
 POP_DENS_BERLIN = 1367  # Population density 2025 in Berlin bounding box per 30 arc second grid cell (from GHS-POP)
 ROAD_LENGTH_PER_CAPITA_BERLIN = 2.1  # Road length per capita in Berlin in meters
+
+MARKET_SHARES = {
+    'petrol_car': 0.534,
+    'diesel_car': 0.25,
+    'motorcycle': 0.082,
+    'bus': 0.001,
+    'rigid_truck': 0.062,
+    'articulated_truck': 0.004,
+    'other': 0.042,
+}
+
+
+class VehicleType(Enum):
+    CAR = 'car'
+    MOTORCYCLE = 'motorcycle'
+    RIGID_TRUCK = 'rigid_truck'
+    ARTICULATED_TRUCK = 'articulated_truck'
+    BUS = 'bus'
+
+
+DENSITY_PETROL = 720
+DENSITY_DIESEL = 820
 
 
 def get_colors_legend(legend_upper_cap: int, color_series: pd.Series) -> tuple[pd.Series, ContinuousLegendData]: