+++ traffic_emissions/core/operator_worker.py
@@ -7,6 +7,10 @@ import shapely
 from climatoology.base.baseoperator import AoiProperties, BaseOperator, ComputationResources, _Artifact
 from climatoology.base.info import _Info
 
+from traffic_emissions.components.traffic_emissions import (
+    get_emission_artifacts,
+    traffic_emissions,
+)
 from traffic_emissions.components.traffic_volume import build_traffic_volume_artifact, traffic_volume
 from traffic_emissions.core.info import get_info
 from traffic_emissions.core.input import ComputeInput
@@ -26,5 +30,12 @@ class Operator(BaseOperator[ComputeInput]):
         params: ComputeInput,
     ) -> List[_Artifact]:
         road_gdf = traffic_volume(aoi)
+        emissions_gdf = traffic_emissions(road_gdf)
+
+        artifacts = []
         traffic_volume_artifact = build_traffic_volume_artifact(road_gdf, resources)
-        return [traffic_volume_artifact]
+        artifacts.append(traffic_volume_artifact)
+        emission_artifacts = get_emission_artifacts(emissions_gdf, resources)
+        artifacts.extend(emission_artifacts)
+
+        return artifacts